:recycle: Refactored /teacher/attend
